Internacionalização

+++ internal/products/data.go
@@ -1,9 +1,9 @@
 package products
 
 type Product struct {
-	ID          uint64 `json:"id"`
+	ID          uint   `json:"id"`
 	Title       string `json:"title"`
 	Description string `json:"description"`
-	Amount      uint64 `json:"amount"`
+	Amount      uint   `json:"amount"`
 	IsGift      bool   `json:"is_gift"`
 }
